requirements.txt
torch==2.1.0
transformers==4.47.0
accelerate==0.27.0
bitsandbytes==0.42.0
datasets==2.20.0
Pillow==10.2.0
numpy==1.26.0
scipy==1.13.0
scikit-learn==1.4.0
matplotlib==3.8.0
seaborn==0.13.0
opencv-python==4.9.0.80
spacy==3.7.0
openai==1.12.0
tqdm==4.66.0
jsonlines==4.0.0
pandas==2.2.0
